requirements.txt
datasets[tabular]==2.18.0
torch==2.2
peft


# BASICS
click
numpy
scipy
pandas
tqdm
click
ipython
# VISUALIZATION
matplotlib
ipympl  # interactive matplotlib plots
seaborn
# altair
# yellowbrick
# ML, STATS & DEEP LEARNING
# statsmodels
# scikit-learn
# sktime
# tslearn
# xgboost
# catboost
# lightgbm
# pytorch
# fastai  # activate fastai channel above!
# tensorflow
# keras
# spacy
# OTHER TOOLS
# optuna
# dask
# snakeviz
tensorboard

optimum

# DEVELOPMENT ONLY PACKAGES (could also be kept in a separate environment file)
pyyaml
tqdm
jupyterlab
ipywidgets
pytest
pytest-cov
tox
pre_commit
nbdime
nbstripout
sphinx
recommonmark
colorlog